Trait pallet_staking_runtime_api::StakingApi

pub trait StakingApi<Block: BlockT, Balance, AccountId>: Core<Block>
where
    Balance: Codec,
    AccountId: Codec,
{
    // Provided methods
    fn nominations_quota(
        &self,
        __runtime_api_at_param__: <Block as BlockT>::Hash,
        balance: Balance,
    ) -> Result<u32, ApiError> { ... }
    fn eras_stakers_page_count(
        &self,
        __runtime_api_at_param__: <Block as BlockT>::Hash,
        era: EraIndex,
        account: AccountId,
    ) -> Result<Page, ApiError> { ... }
    fn pending_rewards(
        &self,
        __runtime_api_at_param__: <Block as BlockT>::Hash,
        era: EraIndex,
        account: AccountId,
    ) -> Result<bool, ApiError> { ... }
}

Provided Methods

fn nominations_quota( &self, __runtime_api_at_param__: <Block as BlockT>::Hash, balance: Balance, ) -> Result<u32, ApiError>

Returns the nominations quota for a nominator with a given balance.

fn eras_stakers_page_count( &self, __runtime_api_at_param__: <Block as BlockT>::Hash, era: EraIndex, account: AccountId, ) -> Result<Page, ApiError>

Returns the page count of exposures for a validator account in a given era.

fn pending_rewards( &self, __runtime_api_at_param__: <Block as BlockT>::Hash, era: EraIndex, account: AccountId, ) -> Result<bool, ApiError>

Returns true if validator account has pages to be claimed for the given era.